# cron/cron-service.go
// Package cron provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package cron

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for ListJobAuditLogsParamsOrder.
const (
	ListJobAuditLogsParamsOrderAsc  ListJobAuditLogsParamsOrder = "asc"
	ListJobAuditLogsParamsOrderDesc ListJobAuditLogsParamsOrder = "desc"
)

// Defines values for ListJobsParamsOrder.
const (
	ListJobsParamsOrderAsc  ListJobsParamsOrder = "asc"
	ListJobsParamsOrderDesc ListJobsParamsOrder = "desc"
)

// Defines values for ListJobsParamsLang.
const (
	ListJobsParamsLangEn ListJobsParamsLang = "en"
	ListJobsParamsLangFr ListJobsParamsLang = "fr"
)

// Defines values for GetJobByIDParamsLang.
const (
	GetJobByIDParamsLangEn GetJobByIDParamsLang = "en"
	GetJobByIDParamsLangFr GetJobByIDParamsLang = "fr"
)

// Defines values for ListRegisteredJobsParamsSortBy.
const (
	IsEnabled        ListRegisteredJobsParamsSortBy = "is_enabled"
	JobName          ListRegisteredJobsParamsSortBy = "job_name"
	LastRegisteredAt ListRegisteredJobsParamsSortBy = "last_registered_at"
	Schedule         ListRegisteredJobsParamsSortBy = "schedule"
)

// Defines values for ListRegisteredJobsParamsOrder.
const (
	ListRegisteredJobsParamsOrderAsc  ListRegisteredJobsParamsOrder = "asc"
	ListRegisteredJobsParamsOrderDesc ListRegisteredJobsParamsOrder = "desc"
)

// Defines values for GetJobAuditLogsParamsOrder.
const (
	GetJobAuditLogsParamsOrderAsc  GetJobAuditLogsParamsOrder = "asc"
	GetJobAuditLogsParamsOrderDesc GetJobAuditLogsParamsOrder = "desc"
)

// ListJobAuditLogsParams defines parameters for ListJobAuditLogs.
type ListJobAuditLogsParams struct {
	// Page page number
	Page *int32 `form:"page,omitempty" json:"page,omitempty"`

	// PageSize maximum number of results to return
	PageSize *int32 `form:"pageSize,omitempty" json:"pageSize,omitempty"`

	// SortBy field to sort by
	SortBy *string `form:"sortBy,omitempty" json:"sortBy,omitempty"`

	// Order sort order
	Order *ListJobAuditLogsParamsOrder `form:"order,omitempty" json:"order,omitempty"`

	// Q starts with
	Q *string `form:"q,omitempty" json:"q,omitempty"`

	// Detail basic or full
	Detail *string `form:"detail,omitempty" json:"detail,omitempty"`
}

// ListJobAuditLogsParamsOrder defines parameters for ListJobAuditLogs.
type ListJobAuditLogsParamsOrder string

// ListJobsParams defines parameters for ListJobs.
type ListJobsParams struct {
	// Page page number
	Page *int32 `form:"page,omitempty" json:"page,omitempty"`

	// PageSize maximum number of results to return
	PageSize *int32 `form:"pageSize,omitempty" json:"pageSize,omitempty"`

	// SortBy field to sort by
	SortBy *string `form:"sortBy,omitempty" json:"sortBy,omitempty"`

	// Order sort order
	Order *ListJobsParamsOrder `form:"order,omitempty" json:"order,omitempty"`

	// Q starts with
	Q *string `form:"q,omitempty" json:"q,omitempty"`

	// Detail basic or full
	Detail *string             `form:"detail,omitempty" json:"detail,omitempty"`
	Lang   *ListJobsParamsLang `form:"lang,omitempty" json:"lang,omitempty"`
}

// ListJobsParamsOrder defines parameters for ListJobs.
type ListJobsParamsOrder string

// ListJobsParamsLang defines parameters for ListJobs.
type ListJobsParamsLang string

// GetJobByIDParams defines parameters for GetJobByID.
type GetJobByIDParams struct {
	Lang *GetJobByIDParamsLang `form:"lang,omitempty" json:"lang,omitempty"`
}

// GetJobByIDParamsLang defines parameters for GetJobByID.
type GetJobByIDParamsLang string

// ListRegisteredJobsParams defines parameters for ListRegisteredJobs.
type ListRegisteredJobsParams struct {
	// Page Page number for pagination
	Page *int32 `form:"page,omitempty" json:"page,omitempty"`

	// PageSize Maximum number of results to return
	PageSize *int32 `form:"pageSize,omitempty" json:"pageSize,omitempty"`

	// SortBy Field to sort by
	SortBy *ListRegisteredJobsParamsSortBy `form:"sortBy,omitempty" json:"sortBy,omitempty"`

	// Order Sort order
	Order *ListRegisteredJobsParamsOrder `form:"order,omitempty" json:"order,omitempty"`

	// Q Search term for job name
	Q *string `form:"q,omitempty" json:"q,omitempty"`
}

// ListRegisteredJobsParamsSortBy defines parameters for ListRegisteredJobs.
type ListRegisteredJobsParamsSortBy string

// ListRegisteredJobsParamsOrder defines parameters for ListRegisteredJobs.
type ListRegisteredJobsParamsOrder string

// UpdateRegisteredJobJSONBody defines parameters for UpdateRegisteredJob.
type UpdateRegisteredJobJSONBody struct {
	// IsEnabled Whether the job is enabled
	IsEnabled *bool `json:"is_enabled,omitempty"`
}

// GetJobAuditLogsParams defines parameters for GetJobAuditLogs.
type GetJobAuditLogsParams struct {
	// Page page number
	Page *int32 `form:"page,omitempty" json:"page,omitempty"`

	// PageSize maximum number of results to return
	PageSize *int32 `form:"pageSize,omitempty" json:"pageSize,omitempty"`

	// SortBy field to sort by
	SortBy *string `form:"sortBy,omitempty" json:"sortBy,omitempty"`

	// Order sort order
	Order *GetJobAuditLogsParamsOrder `form:"order,omitempty" json:"order,omitempty"`

	// Q starts with
	Q *string `form:"q,omitempty" json:"q,omitempty"`
}

// GetJobAuditLogsParamsOrder defines parameters for GetJobAuditLogs.
type GetJobAuditLogsParamsOrder string

// UpdateRegisteredJobJSONRequestBody defines body for UpdateRegisteredJob for application/json ContentType.
type UpdateRegisteredJobJSONRequestBody UpdateRegisteredJobJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /api/v1/cron/job-audit-logs)
	ListJobAuditLogs(c *gin.Context, params ListJobAuditLogsParams)

	// (DELETE /api/v1/cron/job-audit-logs/{id})
	DeleteJobAuditLog(c *gin.Context, id openapi_types.UUID)

	// (GET /api/v1/cron/job-audit-logs/{id})
	GetJobAuditLogByID(c *gin.Context, id openapi_types.UUID)

	// (GET /api/v1/cron/jobs)
	ListJobs(c *gin.Context, params ListJobsParams)

	// (DELETE /api/v1/cron/jobs/{id})
	DeleteJob(c *gin.Context, id openapi_types.UUID)

	// (GET /api/v1/cron/jobs/{id})
	GetJobByID(c *gin.Context, id openapi_types.UUID, params GetJobByIDParams)

	// (POST /api/v1/cron/migrate/down)
	MigrateDown(c *gin.Context)

	// (POST /api/v1/cron/migrate/up)
	MigrateUp(c *gin.Context)

	// (GET /api/v1/cron/registered-jobs)
	ListRegisteredJobs(c *gin.Context, params ListRegisteredJobsParams)

	// (GET /api/v1/cron/registered-jobs/{id})
	GetRegisteredJob(c *gin.Context, id openapi_types.UUID)

	// (PATCH /api/v1/cron/registered-jobs/{id})
	UpdateRegisteredJob(c *gin.Context, id openapi_types.UUID)

	// (GET /api/v1/cron/registered-jobs/{id}/audit-logs)
	GetJobAuditLogs(c *gin.Context, id openapi_types.UUID, params GetJobAuditLogsParams)

	// (POST /api/v1/cron/seed/reference)
	SeedReferenceData(c *gin.Context)

	// (POST /api/v1/cron/seed/sample)
	SeedSampleData(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// ListJobAuditLogs operation middleware
func (siw *ServerInterfaceWrapper) ListJobAuditLogs(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListJobAuditLogsParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", c.Request.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter page: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "pageSize" -------------

	err = runtime.BindQueryParameter("form", true, false, "pageSize", c.Request.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter pageSize: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "sortBy" -------------

	err = runtime.BindQueryParameter("form", true, false, "sortBy", c.Request.URL.Query(), &params.SortBy)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sortBy: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "order" -------------

	err = runtime.BindQueryParameter("form", true, false, "order", c.Request.URL.Query(), &params.Order)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter order: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "q" -------------

	err = runtime.BindQueryParameter("form", true, false, "q", c.Request.URL.Query(), &params.Q)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter q: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "detail" -------------

	err = runtime.BindQueryParameter("form", true, false, "detail", c.Request.URL.Query(), &params.Detail)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter detail: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListJobAuditLogs(c, params)
}

// DeleteJobAuditLog operation middleware
func (siw *ServerInterfaceWrapper) DeleteJobAuditLog(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteJobAuditLog(c, id)
}

// GetJobAuditLogByID operation middleware
func (siw *ServerInterfaceWrapper) GetJobAuditLogByID(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetJobAuditLogByID(c, id)
}

// ListJobs operation middleware
func (siw *ServerInterfaceWrapper) ListJobs(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListJobsParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", c.Request.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter page: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "pageSize" -------------

	err = runtime.BindQueryParameter("form", true, false, "pageSize", c.Request.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter pageSize: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "sortBy" -------------

	err = runtime.BindQueryParameter("form", true, false, "sortBy", c.Request.URL.Query(), &params.SortBy)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sortBy: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "order" -------------

	err = runtime.BindQueryParameter("form", true, false, "order", c.Request.URL.Query(), &params.Order)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter order: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "q" -------------

	err = runtime.BindQueryParameter("form", true, false, "q", c.Request.URL.Query(), &params.Q)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter q: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "detail" -------------

	err = runtime.BindQueryParameter("form", true, false, "detail", c.Request.URL.Query(), &params.Detail)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter detail: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "lang" -------------

	err = runtime.BindQueryParameter("form", true, false, "lang", c.Request.URL.Query(), &params.Lang)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter lang: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListJobs(c, params)
}

// DeleteJob operation middleware
func (siw *ServerInterfaceWrapper) DeleteJob(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteJob(c, id)
}

// GetJobByID operation middleware
func (siw *ServerInterfaceWrapper) GetJobByID(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetJobByIDParams

	// ------------- Optional query parameter "lang" -------------

	err = runtime.BindQueryParameter("form", true, false, "lang", c.Request.URL.Query(), &params.Lang)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter lang: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetJobByID(c, id, params)
}

// MigrateDown operation middleware
func (siw *ServerInterfaceWrapper) MigrateDown(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.MigrateDown(c)
}

// MigrateUp operation middleware
func (siw *ServerInterfaceWrapper) MigrateUp(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.MigrateUp(c)
}

// ListRegisteredJobs operation middleware
func (siw *ServerInterfaceWrapper) ListRegisteredJobs(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListRegisteredJobsParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", c.Request.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter page: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "pageSize" -------------

	err = runtime.BindQueryParameter("form", true, false, "pageSize", c.Request.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter pageSize: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "sortBy" -------------

	err = runtime.BindQueryParameter("form", true, false, "sortBy", c.Request.URL.Query(), &params.SortBy)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sortBy: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "order" -------------

	err = runtime.BindQueryParameter("form", true, false, "order", c.Request.URL.Query(), &params.Order)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter order: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "q" -------------

	err = runtime.BindQueryParameter("form", true, false, "q", c.Request.URL.Query(), &params.Q)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter q: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListRegisteredJobs(c, params)
}

// GetRegisteredJob operation middleware
func (siw *ServerInterfaceWrapper) GetRegisteredJob(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetRegisteredJob(c, id)
}

// UpdateRegisteredJob operation middleware
func (siw *ServerInterfaceWrapper) UpdateRegisteredJob(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateRegisteredJob(c, id)
}

// GetJobAuditLogs operation middleware
func (siw *ServerInterfaceWrapper) GetJobAuditLogs(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetJobAuditLogsParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", c.Request.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter page: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "pageSize" -------------

	err = runtime.BindQueryParameter("form", true, false, "pageSize", c.Request.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter pageSize: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "sortBy" -------------

	err = runtime.BindQueryParameter("form", true, false, "sortBy", c.Request.URL.Query(), &params.SortBy)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sortBy: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "order" -------------

	err = runtime.BindQueryParameter("form", true, false, "order", c.Request.URL.Query(), &params.Order)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter order: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Optional query parameter "q" -------------

	err = runtime.BindQueryParameter("form", true, false, "q", c.Request.URL.Query(), &params.Q)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter q: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetJobAuditLogs(c, id, params)
}

// SeedReferenceData operation middleware
func (siw *ServerInterfaceWrapper) SeedReferenceData(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.SeedReferenceData(c)
}

// SeedSampleData operation middleware
func (siw *ServerInterfaceWrapper) SeedSampleData(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.SeedSampleData(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/api/v1/cron/job-audit-logs", wrapper.ListJobAuditLogs)
	router.DELETE(options.BaseURL+"/api/v1/cron/job-audit-logs/:id", wrapper.DeleteJobAuditLog)
	router.GET(options.BaseURL+"/api/v1/cron/job-audit-logs/:id", wrapper.GetJobAuditLogByID)
	router.GET(options.BaseURL+"/api/v1/cron/jobs", wrapper.ListJobs)
	router.DELETE(options.BaseURL+"/api/v1/cron/jobs/:id", wrapper.DeleteJob)
	router.GET(options.BaseURL+"/api/v1/cron/jobs/:id", wrapper.GetJobByID)
	router.POST(options.BaseURL+"/api/v1/cron/migrate/down", wrapper.MigrateDown)
	router.POST(options.BaseURL+"/api/v1/cron/migrate/up", wrapper.MigrateUp)
	router.GET(options.BaseURL+"/api/v1/cron/registered-jobs", wrapper.ListRegisteredJobs)
	router.GET(options.BaseURL+"/api/v1/cron/registered-jobs/:id", wrapper.GetRegisteredJob)
	router.PATCH(options.BaseURL+"/api/v1/cron/registered-jobs/:id", wrapper.UpdateRegisteredJob)
	router.GET(options.BaseURL+"/api/v1/cron/registered-jobs/:id/audit-logs", wrapper.GetJobAuditLogs)
	router.POST(options.BaseURL+"/api/v1/cron/seed/reference", wrapper.SeedReferenceData)
	router.POST(options.BaseURL+"/api/v1/cron/seed/sample", wrapper.SeedSampleData)
}